[lab06] adr adrp
1. test adr
2. test adrp
