+++ lab06/gdb.sh
